Implments the Roman Numeral Kata
So, as my first attempt at trying out some Elixir (beyond playing around
in iex), I implemented the Roman Numeral kata.

This first vertion works, but feels like it's using a lot of brute force
repitition. There's probably a more elegent way to accomplish it.

In any case, the test are all passing, so I'm calling a it a good first
attempt.