Implements Fibonacci module
As part of the Elixir study group at RF, we implemented the Project
Euler problem Two, "Even Fibonacci Numbers":

> Each new term in the Fibonacci sequence is generated by adding
> the previous two terms. By starting with 1 and 2, the first 10 terms
> will be:
>
> 1, 2, 3, 5, 8, 13, 21, 34, 55, 89, ...
>
> By considering the terms in the Fibonacci sequence whose values do not
> exceed four million, find the sum of the even-valued terms.

See: https://projecteuler.net/problem=2

---

To solve the problem, I created a module, Fib with two methods:

fib/1

and

sum_of_evens/1

The fib/1 function calculates the n-th number in the Fibonacci sequence.

The sum_of_evens_to/1 calculates the sum of the even numbers in the
sequence up to a limit. (Within the first 1 million numbers.)

Tests: all green

Answer to Project Euler question: 4,613,732
(No idea if this is correct, but the answers are correct for smaller
number I calculated by hand.)