Another pass at the Fib katas
This time around the implementation of the "sum of evens"
kata was significantly bigger, and arguably more complex.
But the old, simple version had a performance issue
in which building the list from which the sum was derived
involved building the fib sequence over and over and over
again. Building a sequence of n elements required nearly n^2 * 3
calculations.

The new implementation builds up a single instance of the
Fib sequence to the limit, then just piping it through a filter
& sum function. More code, but more efficient.