Refactor Fib kata to use Stream
Replaced fib implementation with methods
that use a Stream of sequence.

Vastly simplified the code:

Reduced LOC to less than a third of prev impl.

sequence_of/1 and element/1 each became one-liners.

sum_of_evens_to/1 is a bit bigger but still pretty
simple. The trick was to pipe the stream to
Enum.reduce_while/3. The function in reduce_while needs
to return a tuple, {:cont, acc} or {:halt, acc}. When it
gets the {:halt, acc}, the acc value is returned.