WIP: 'Enum Functions' exercise - each/1
Implements each/1 function.

Not sure how to test this because the return is always :ok.
Code in my_list_test doen't work. Tried using an IO.puts/1 w/
ExUnit.CaptureIO, but it isn't working because it feeds the
:ok result to capture_io/1 along with the stdout capture. There's
probably some simple workaround for this, but I can't think of it
at the moment.

Doctest works and is passing (though the screen output is annoying).